+++ ext/fluxnet_simulations/US-Ha1.jl
@@ -120,20 +120,12 @@ function FluxnetSimulations.get_parameters(
     ψ63 = FT(-4 / 0.0098),
     Weibull_param = FT(4),
     a = FT(0.05 * 0.0098),
-    conductivity_model = PlantHydraulics.Weibull{FT}(
-        K_sat_plant,
-        ψ63,
-        Weibull_param,
-    ),
-    retention_model = PlantHydraulics.LinearRetentionCurve{FT}(a),
+    conductivity_model = Canopy.Weibull{FT}(K_sat_plant, ψ63, Weibull_param),
+    retention_model = Canopy.LinearRetentionCurve{FT}(a),
     plant_ν = FT(2.46e-4),
     plant_S_s = FT(1e-2 * 0.0098),
     rooting_depth = FT(0.5),
-    n_stem = Int64(1),
-    n_leaf = Int64(1),
-    h_leaf = FT(12),
-    h_stem = FT(14),
-    h_canopy = h_leaf + h_stem,
+    h_canopy = FT(26),
 )
     return (;
         soil_ν,
@@ -175,10 +167,6 @@ function FluxnetSimulations.get_parameters(
         plant_ν,
         plant_S_s,
         rooting_depth,
-        n_stem,
-        n_leaf,
-        h_leaf,
-        h_stem,
         h_canopy,
     )